Create utils.py
Step 10: Refactoring utils.py

Create the File: Make sure you have created an empty file named utils.py.

Analyze Code: See if there is any code which doesn't fit within other modules. There are no methods we have explicitly defined in the other modules which do not belong, so this file is mostly empty for now.

Helper functions for future: You can use this file to define functions or methods that you use in multiple parts of the program. For now, we can leave this blank, and we can add utility methods to it as we continue working on it.

Test: Since we didn't add anything, there shouldn't be any problems with testing.